requirements.txt
absl-py==0.13.0
astor==0.7.1
gast==0.2.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.5
numpy==1.17.4
Pillow==5.4.1
protobuf==3.6.1
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
termcolor==1.1.0
